Improve the method of initialize
